+++ src/index.tsx
@@ -1,6 +1,6 @@
+import App from "components/app/App";
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
 import "./index.css";
 import * as serviceWorker from "./serviceWorker";
